# Add `Clone` constraint for `Model::Action` and simplify next_steps