Instead of calling `Module::actions` twice in `Module::next_steps`, it is better to clone the result directly. Because `Model::actions` can sometimes be expensive.