@jonnadal 2023-05-28
Hi @HaoYang670. The repeated call to `Model::actions` is intentional to avoid putting a `Clone` constraint on `Model::Actions`. `Model::next_steps` is only called from performance insensitive code (not in any of the central loops) so that felt like a better tradeoff.

I'm open to reevaluating the decision if you feel strongly about this change, but overall I lean towards keeping the type constraints minimal unless there's an offsetting benefit to users of the library. Please let me know what you think. I'm also interested in others' opinions if anyone else notices this PR.