@HaoYang670 2023-05-29
Thank you for the quick response @jonnadal 

I did benchmarks on the 2 branchs, and found no obvious performance imporvement (you are right, it is not performance sensitive):
```
master branch:

== 2pc check 10 ==
Done. states=817760258, unique=61515776, depth=32, sec=26
Done. states=817760258, unique=61515776, depth=32, sec=26
Done. states=817760258, unique=61515776, depth=32, sec=25
== paxos check 6 ==
Done. states=18972647, unique=9357525, depth=28, sec=6
Done. states=18972647, unique=9357525, depth=28, sec=6
Done. states=18972647, unique=9357525, depth=28, sec=6
== single-copy-register check 4 ==
Done. states=731789, unique=400233, depth=17, sec=2
Done. states=731789, unique=400233, depth=17, sec=3
Done. states=731789, unique=400233, depth=17, sec=3
== linearizable-register check 2 ==
Done. states=194646243, unique=41138199, depth=33, sec=28
Done. states=194646243, unique=41138199, depth=33, sec=28
Done. states=194646243, unique=41138199, depth=33, sec=28
== linearizable-register check 3 ordered ==

this branch

== 2pc check 10 ==
Done. states=817760258, unique=61515776, depth=32, sec=27
Done. states=817760258, unique=61515776, depth=32, sec=27
Done. states=817760258, unique=61515776, depth=32, sec=25
== paxos check 6 ==
Done. states=18972647, unique=9357525, depth=28, sec=6
Done. states=18972647, unique=9357525, depth=28, sec=6
Done. states=18972647, unique=9357525, depth=28, sec=6
== single-copy-register check 4 ==
Done. states=731789, unique=400233, depth=17, sec=3
Done. states=731789, unique=400233, depth=17, sec=3
Done. states=731789, unique=400233, depth=17, sec=3
== linearizable-register check 2 ==
Done. states=194646243, unique=41138199, depth=33, sec=28
Done. states=194646243, unique=41138199, depth=33, sec=27
Done. states=194646243, unique=41138199, depth=33, sec=28

```